Separate enum ModelPacketType for CRUD operation identification
