+++ src/main/java/net/packet/model/ModelPacketType.java
@@ -1,0 +1,16 @@
+package net.packet.model;
+
+public enum ModelPacketType {
+
+    INVALID(-1), CREATE(0), UPDATE_LIBBOOK(2), DELETE(3);
+
+    private int typeId;
+
+    ModelPacketType(int typeId) {
+        this.typeId = typeId;
+    }
+
+    public int getTypeId() {
+        return typeId;
+    }
+}
