+++ src/main/java/net/packet/model/book/LibraryBookPacket.java
@@ -1,0 +1,26 @@
+package net.packet.model.book;
+
+import model.item.LibraryBook;
+import net.packet.Packet;
+import net.packet.model.ModelPacketType;
+
+import java.io.Serializable;
+
+public class LibraryBookPacket extends Packet implements Serializable {
+
+    private ModelPacketType packetType;
+    private LibraryBook book;
+
+    public LibraryBookPacket(ModelPacketType packetType, LibraryBook book) {
+        this.packetType = packetType;
+        this.book = book;
+    }
+
+    public ModelPacketType getPacketType() {
+        return packetType;
+    }
+
+    public LibraryBook getBook() {
+        return book;
+    }
+}
